refactor(ui): extract proxy base-URL and header resolution into src/lib/http
Moves the stateful API-target resolution out of the networking.tsx god-module
into src/lib/http/proxyBase.ts: proxyBaseUrl, serverRootPath, the worker-URL
localStorage handling, getProxyBaseUrl, updateProxyBaseUrl, switchToWorkerUrl,
getWindowLocation, and the global header name get/set. networking.tsx now imports
what it uses and re-exports the previously-public symbols, so its ~300 importers
are untouched.

This flips the one backwards dependency the http layer had: src/lib/http no longer
needs to reach into the component tree for the base URL, which is what was forcing
the new openapi-fetch client (api.ts) to import from networking. It's a pure
relocation; the singleton state is shared because one module instance is imported
by both networking and the http client. Drops the dead updateServerRootPath, and
replaces the isLocal console.log silencing with an inline NODE_ENV check.

Adds proxyBase.test.ts covering the getProxyBaseUrl window-origin fallback, the
worker-URL switch and its non-http-scheme rejection, and the header-name get/set.
